+++ src/ogl/glArchivItem_Font.h
@@ -45,8 +45,8 @@ class glArchivItem_Font : public libsiedler2::ArchivItem_Font
         void Draw(short x, short y, const std::string& text,   unsigned int format, unsigned int color = COLOR_WHITE, unsigned short length = 0, unsigned short max = 0xFFFF, const std::string& end   = "...",  unsigned short end_length = 0);
 
         /// liefert die Länge einer Zeichenkette.
-        unsigned short getWidth(const std::wstring& text, unsigned length = 0, unsigned max_width = 0xffffffff, unsigned short* max = NULL) const;
-        unsigned short getWidth(const std::string& text, unsigned length = 0, unsigned max_width = 0xffffffff, unsigned short* max = NULL) const;
+        unsigned short getWidth(const std::wstring& text, unsigned length = 0, unsigned max_width = 0xffffffff, unsigned* max = NULL) const;
+        unsigned short getWidth(const std::string& text, unsigned length = 0, unsigned max_width = 0xffffffff, unsigned* max = NULL) const;
         /// liefert die Höhe des Textes ( entspricht @p getDy()+1 )
         inline unsigned short getHeight() const { return dy + 1; }
